test: isolate SessionDO eviction coverage
