+++ packages/control-plane/test/integration/helpers.ts
@@ -8,6 +8,7 @@ import { SessionIndexStore } from "../../src/db/session-index";
 import type { SessionModelProviderAuthInput } from "../../src/model-provider-accounts/provider-auth-contracts";
 
 const DEFAULT_WAIT_FOR_SANDBOX_STATUS_TIMEOUT_MS = 3000;
+export const INTEGRATION_WEBSOCKET_TIMEOUT_MS = 2000;
 const TEST_BROWSER_USER_ID = "11111111111111111111111111111111";
 const TEST_BROWSER_ACCOUNT_ID = "test-browser-account";
 const TEST_BROWSER_PROVIDER_SUBJECT = "583231";
@@ -375,7 +376,7 @@ export function collectMessages(
 ): Promise<Record<string, unknown>[]> {
   return new Promise((resolve) => {
     const messages: Record<string, unknown>[] = [];
-    const timeout = opts?.timeoutMs ?? 2000;
+    const timeout = opts?.timeoutMs ?? INTEGRATION_WEBSOCKET_TIMEOUT_MS;
     const timer = setTimeout(() => resolve(messages), timeout);
 
     ws.addEventListener("message", (event) => {
